Ignore missing directories

+++ toolbox.js
@@ -25,19 +25,24 @@
         fCallback = typeof fCallback == 'function' ? fCallback  : function() {};
 
         fs.stat(sPath, function(oError, oStat) {
-            if (oStat.isDirectory()) {
-                exec('rm ' + path.join(sPath, '/*'), function() {
-                    console.log('removed', path.join(sPath, '/*'));
-                    fs.rmdir(sPath, function() {
+            if (oStat !== undefined) {
+                if (oStat.isDirectory()) {
+                    exec('rm ' + path.join(sPath, '/*'), function() {
+                        console.log('removed', path.join(sPath, '/*'));
+                        fs.rmdir(sPath, function() {
+                            console.log('removed', sPath);
+                            fCallback(sPath);
+                        });
+                    });
+                } else {
+                    fs.unlink(sPath, function() {
                         console.log('removed', sPath);
                         fCallback(sPath);
                     });
-                });
+                }
             } else {
-                fs.unlink(sPath, function() {
-                    console.log('removed', sPath);
-                    fCallback(sPath);
-                });
+                console.log('did not find for removal', sPath);
+                fCallback(sPath);
             }
         });
     };
